Samle apper som videresender forespoersel-eventer
