+++ apps/integrasjonstest/build.gradle.kts
@@ -12,11 +12,7 @@ val apps =
         "db",
         "distribusjon",
         "feil-behandler",
-        "forespoersel-besvart",
-        "forespoersel-forkastet",
-        "forespoersel-infotrygd",
         "forespoersel-marker-besvart",
-        "forespoersel-mottatt",
         "helsebro",
         "innsending",
         "inntekt",
