--- config/forespoersel-forkastet/prod-gcp.yml
@@ -1,4 +1,0 @@
-kafkaPool: nav-prod
-env:
-  - name: KAFKA_EXTRA_TOPIC
-    value: helsearbeidsgiver.pri
