--- config/forespoersel-mottatt/dev-gcp.yml
@@ -1,4 +1,0 @@
-kafkaPool: nav-dev
-env:
-  - name: KAFKA_EXTRA_TOPIC
-    value: helsearbeidsgiver.pri
